// src/api/api.ts
export const BASE_PATH = 'http://peterserver1.myvnc.com:8080/rest'.replace(/\/+$/, '');

/**
 *
 * @export
 */
export const COLLECTION_FORMATS = {
  csv: ',',
  ssv: ' ',
  tsv: '\t',
  pipes: '|'
};

/**
 *
 * @export
 * @interface FetchAPI
 */
export interface FetchAPI {
  (url: string, init?: any): Promise<Response>;
}

/**
 *
 * @export
 * @interface FetchArgs
 */
export interface FetchArgs {
  url: string;
  options: any;
}

/**
 *
 * @export
 * @class BaseAPI
 */
export class BaseAPI {
  protected configuration: Configuration = new Configuration();

  constructor(
    configuration?: Configuration,
    protected basePath: string = BASE_PATH,
    protected fetch: FetchAPI = portableFetch
  ) {
    if (configuration) {
      this.configuration = configuration;
      this.basePath = configuration.basePath || this.basePath;
    }
  }
}
// ...
export interface CommandDescription {
  /**
   *
   * @type {Array<CommandOption>}
   * @memberof CommandDescription
   */
  commandOptions?: Array<CommandOption>;
}
// ...
export interface EnrichedItemDTO {
  /**
   *
   * @type {string}
   * @memberof EnrichedItemDTO
   */
  type?: string;
  /**
   *
   * @type {string}
   * @memberof EnrichedItemDTO
   */
  name?: string;
  /**
   *
   * @type {string}
   * @memberof EnrichedItemDTO
   */
  label?: string;
  /**
   *
   * @type {string}
   * @memberof EnrichedItemDTO
   */
  category?: string;
  /**
   *
   * @type {Array<string>}
   * @memberof EnrichedItemDTO
   */
  tags?: Array<string>;
  /**
   *
   * @type {Array<string>}
   * @memberof EnrichedItemDTO
   */
  groupNames?: Array<string>;
  /**
   *
   * @type {string}
   * @memberof EnrichedItemDTO
   */
  link?: string;
  /**
   *
   * @type {string}
   * @memberof EnrichedItemDTO
   */
  state?: string;
  /**
   *
   * @type {string}
   * @memberof EnrichedItemDTO
   */
  transformedState?: string;
  /**
   *
   * @type {StateDescription}
   * @memberof EnrichedItemDTO
   */
  stateDescription?: StateDescription;
  /**
   *
   * @type {CommandDescription}
   * @memberof EnrichedItemDTO
   */
  commandDescription?: CommandDescription;
  /**
   *
   * @type {{ [key: string]: any; }}
   * @memberof EnrichedItemDTO
   */
  metadata?: { [key: string]: any };
  /**
   *
   * @type {boolean}
   * @memberof EnrichedItemDTO
   */
  editable?: boolean;
}
// ...
export interface MappingDTO {
  /**
   *
   * @type {string}
   * @memberof MappingDTO
   */
  command?: string;
  /**
   *
   * @type {string}
   * @memberof MappingDTO
   */
  label?: string;
}

/**
 *
 * @export
 * @interface MetadataDTO
 */
export interface MetadataDTO {
  /**
   *
   * @type {string}
   * @memberof MetadataDTO
   */
  value?: string;
  /**
   *
   * @type {{ [key: string]: any; }}
   * @memberof MetadataDTO
   */
  config?: { [key: string]: any };
}

/**
 *
 * @export
 * @interface PageDTO
 */
export interface PageDTO {
  /**
   *
   * @type {string}
   * @memberof PageDTO
   */
  id?: string;
  /**
   *
   * @type {string}
   * @memberof PageDTO
   */
  title?: string;
  /**
   *
   * @type {string}
   * @memberof PageDTO
   */
  icon?: string;
  /**
   *
   * @type {string}
   * @memberof PageDTO
   */
  link?: string;
  /**
   *
   * @type {PageDTO}
   * @memberof PageDTO
   */
  parent?: PageDTO;
  /**
   *
   * @type {boolean}
   * @memberof PageDTO
   */
  leaf?: boolean;
  /**
   *
   * @type {boolean}
   * @memberof PageDTO
   */
  timeout?: boolean;
  /**
   *
   * @type {Array<WidgetDTO>}
   * @memberof PageDTO
   */
  widgets?: Array<WidgetDTO>;
}
// ...
export interface SitemapDTO {
  /**
   *
   * @type {string}
   * @memberof SitemapDTO
   */
  name?: string;
  /**
   *
   * @type {string}
   * @memberof SitemapDTO
   */
  icon?: string;
  /**
   *
   * @type {string}
   * @memberof SitemapDTO
   */
  label?: string;
  /**
   *
   * @type {string}
   * @memberof SitemapDTO
   */
  link?: string;
  /**
   *
   * @type {PageDTO}
   * @memberof SitemapDTO
   */
  homepage?: PageDTO;
}

/**
 *
 * @export
 * @interface StateDescription
 */
export interface StateDescription {
  /**
   *
   * @type {number}
   * @memberof StateDescription
   */
  minimum?: number;
  /**
   *
   * @type {number}
   * @memberof StateDescription
   */
  maximum?: number;
  /**
   *
   * @type {number}
   * @memberof StateDescription
   */
  step?: number;
  /**
   *
   * @type {string}
   * @memberof StateDescription
   */
  pattern?: string;
  /**
   *
   * @type {boolean}
   * @memberof StateDescription
   */
  readOnly?: boolean;
  /**
   *
   * @type {Array<StateOption>}
   * @memberof StateDescription
   */
  options?: Array<StateOption>;
}
// ...
export interface WidgetDTO {
  /**
   *
   * @type {string}
   * @memberof WidgetDTO
   */
  widgetId?: string;
  /**
   *
   * @type {string}
   * @memberof WidgetDTO
   */
  type?: string;
  /**
   *
   * @type {string}
   * @memberof WidgetDTO
   */
  name?: string;
  /**
   *
   * @type {boolean}
   * @memberof WidgetDTO
   */
  visibility?: boolean;
  /**
   *
   * @type {string}
   * @memberof WidgetDTO
   */
  label?: string;
  /**
   *
   * @type {string}
   * @memberof WidgetDTO
   */
  icon?: string;
  /**
   *
   * @type {string}
   * @memberof WidgetDTO
   */
  labelcolor?: string;
  /**
   *
   * @type {string}
   * @memberof WidgetDTO
   */
  valuecolor?: string;
  /**
   *
   * @type {Array<MappingDTO>}
   * @memberof WidgetDTO
   */
  mappings?: Array<MappingDTO>;
  /**
   *
   * @type {boolean}
   * @memberof WidgetDTO
   */
  switchSupport?: boolean;
  /**
   *
   * @type {number}
   * @memberof WidgetDTO
   */
  sendFrequency?: number;
  /**
   *
   * @type {string}
   * @memberof WidgetDTO
   */
  separator?: string;
  /**
   *
   * @type {number}
   * @memberof WidgetDTO
   */
  refresh?: number;
  /**
   *
   * @type {number}
   * @memberof WidgetDTO
   */
  height?: number;
  /**
   *
   * @type {number}
   * @memberof WidgetDTO
   */
  minValue?: number;
  /**
   *
   * @type {number}
   * @memberof WidgetDTO
   */
  maxValue?: number;
  /**
   *
   * @type {number}
   * @memberof WidgetDTO
   */
  step?: number;
  /**
   *
   * @type {string}
   * @memberof WidgetDTO
   */
  url?: string;
  /**
   *
   * @type {string}
   * @memberof WidgetDTO
   */
  encoding?: string;
  /**
   *
   * @type {string}
   * @memberof WidgetDTO
   */
  service?: string;
  /**
   *
   * @type {string}
   * @memberof WidgetDTO
   */
  period?: string;
  /**
   *
   * @type {boolean}
   * @memberof WidgetDTO
   */
  legend?: boolean;
  /**
   *
   * @type {string}
   * @memberof WidgetDTO
   */
  state?: string;
  /**
   *
   * @type {EnrichedItemDTO}
   * @memberof WidgetDTO
   */
  item?: EnrichedItemDTO;
  /**
   *
   * @type {PageDTO}
   * @memberof WidgetDTO
   */
  linkedPage?: PageDTO;
  /**
   *
   * @type {Array<WidgetDTO>}
   * @memberof WidgetDTO
   */
  widgets?: Array<WidgetDTO>;
}
// ...
export interface SitemapsApiInterface {
  /**
   *
   * @summary Creates a sitemap event subscription.
   * @param {*} [options] Override http request option.
   * @throws {RequiredError}
   * @memberof SitemapsApiInterface
   */
  createEventSubscription(options?: any): Promise<any>;

  /**
   *
   * @summary Polls the data for a sitemap.
   * @param {string} sitemapname sitemap name
   * @param {string} pageid page id
   * @param {string} [acceptLanguage] language
   * @param {string} [subscriptionid] subscriptionid
   * @param {boolean} [includeHidden] include hidden widgets
   * @param {*} [options] Override http request option.
   * @throws {RequiredError}
   * @memberof SitemapsApiInterface
   */
  getPageData(
    sitemapname: string,
    pageid: string,
    acceptLanguage?: string,
    subscriptionid?: string,
    includeHidden?: boolean,
    options?: any
  ): Promise<{}>;

  /**
   *
   * @summary Get sitemap by name.
   * @param {string} sitemapname sitemap name
   * @param {string} [acceptLanguage] language
   * @param {string} [type]
   * @param {string} [jsoncallback]
   * @param {boolean} [includeHidden] include hidden widgets
   * @param {*} [options] Override http request option.
   * @throws {RequiredError}
   * @memberof SitemapsApiInterface
   */
  getSitemapData(
    sitemapname: string,
    acceptLanguage?: string,
    type?: string,
    jsoncallback?: string,
    includeHidden?: boolean,
    options?: any
  ): Promise<{}>;

  /**
   *
   * @summary Get sitemap events.
   * @param {string} subscriptionid subscription id
   * @param {string} [sitemap] sitemap name
   * @param {string} [pageid] page id
   * @param {*} [options] Override http request option.
   * @throws {RequiredError}
   * @memberof SitemapsApiInterface
   */
  getSitemapEvents(
    subscriptionid: string,
    sitemap?: string,
    pageid?: string,
    options?: any
  ): Promise<{}>;

  /**
   *
   * @summary Get all available sitemaps.
   * @param {*} [options] Override http request option.
   * @throws {RequiredError}
   * @memberof SitemapsApiInterface
   */
  getSitemaps(options?: any): Promise<{}>;
}

/**
 * SitemapsApi - object-oriented interface
 * @export
 * @class SitemapsApi
 * @extends {BaseAPI}
 */
export class SitemapsApi extends BaseAPI implements SitemapsApiInterface {
  /**
   *
   * @summary Creates a sitemap event subscription.
   * @param {*} [options] Override http request option.
   * @throws {RequiredError}
   * @memberof SitemapsApi
   */
  public createEventSubscription(options?: any) {
    return SitemapsApiFp(this.configuration).createEventSubscription(options)(
      this.fetch,
      this.basePath
    );
  }

  /**
   *
   * @summary Polls the data for a sitemap.
   * @param {string} sitemapname sitemap name
   * @param {string} pageid page id
   * @param {string} [acceptLanguage] language
   * @param {string} [subscriptionid] subscriptionid
   * @param {boolean} [includeHidden] include hidden widgets
   * @param {*} [options] Override http request option.
   * @throws {RequiredError}
   * @memberof SitemapsApi
   */
  public getPageData(
    sitemapname: string,
    pageid: string,
    acceptLanguage?: string,
    subscriptionid?: string,
    includeHidden?: boolean,
    options?: any
  ) {
    return SitemapsApiFp(this.configuration).getPageData(
      sitemapname,
      pageid,
      acceptLanguage,
      subscriptionid,
      includeHidden,
      options
    )(this.fetch, this.basePath);
  }

  /**
   *
   * @summary Get sitemap by name.
   * @param {string} sitemapname sitemap name
   * @param {string} [acceptLanguage] language
   * @param {string} [type]
   * @param {string} [jsoncallback]
   * @param {boolean} [includeHidden] include hidden widgets
   * @param {*} [options] Override http request option.
   * @throws {RequiredError}
   * @memberof SitemapsApi
   */
  public getSitemapData(
    sitemapname: string,
    acceptLanguage?: string,
    type?: string,
    jsoncallback?: string,
    includeHidden?: boolean,
    options?: any
  ) {
    return SitemapsApiFp(this.configuration).getSitemapData(
      sitemapname,
      acceptLanguage,
      type,
      jsoncallback,
      includeHidden,
      options
    )(this.fetch, this.basePath);
  }

  /**
   *
   * @summary Get sitemap events.
   * @param {string} subscriptionid subscription id
   * @param {string} [sitemap] sitemap name
   * @param {string} [pageid] page id
   * @param {*} [options] Override http request option.
   * @throws {RequiredError}
   * @memberof SitemapsApi
   */
  public getSitemapEvents(
    subscriptionid: string,
    sitemap?: string,
    pageid?: string,
    options?: any
  ) {
    return SitemapsApiFp(this.configuration).getSitemapEvents(
      subscriptionid,
      sitemap,
      pageid,
      options
    )(this.fetch, this.basePath);
  }

  /**
   *
   * @summary Get all available sitemaps.
   * @param {*} [options] Override http request option.
   * @throws {RequiredError}
   * @memberof SitemapsApi
   */
  public getSitemaps(options?: any) {
    return SitemapsApiFp(this.configuration).getSitemaps(options)(this.fetch, this.basePath);
  }
}